automatically cancel focus time when the computer is locked
This feature is supposed to help users avoid keeping their tracker on
accidentally when they step away from their computer. But it could also
be sensibly used as a feature that let's users simply press the lock
button on their computer and get up to make the most of their break.