improve behavior on idle detection
First, the idle time should start not only when the machine switches to
the idle state, as the user has been idle already for the configured
period. Account for that.

Second, when a user is idle for long enough, so that the screen is
automatically locked, the lock behavior would override the idle behavior
of the focus tracker. So a locked state should cancel the focus timer
only when there is no pending idle notification.