make history operations update the total focus time
This is achieved by keeping track of the history entries that went into
the total focus time. So when an element is removed or updated, we can
tell whether that affects the total tracked time.

For manually added entries, we assume they are supposed to pertain to
the current session and, thus, should go into the subtotal.