allow to write up notes before a focus event has completed
Instead of going to the sidebar and editing existing focus events, users
can write what they're working on in the popup and it will be
automatically annotated to all future focus events.

Not the best UX: You have to press Enter to save the notes. A
light-orange background signals when a change has not been stored yet.