fix incorrect initial history date
Turns out that `new Date()` interprets date-only strings without an
explicit timezone as UTC where date-time strings without a timezone are
interpreted in the local timezone. As a result, the initial value of
`historyDate` would not be set to the start of the day unless the local
timezone happens to be `+0000`.