incorporate pause time prior preceding ongoing focus time
The code so far missed to add pause times after the latested completed
focus time and a started-yet-not-finished focus time to the total.
Hence, the pause total tended to be underreported while in focus time.

Besides updating the pause total in the popup, another row is added to
the history table for ongoing focus times to display the pause time
preceeding it.

There's also some refactoring done here that touches the updated code.