allow to resume the previous focus time
When a focus time has been incorrectly left, users could start a new
focus time, but it would clutter the history and possibly leave a gap in
the tracked time.

This adds a drop-down menu to the focus button, so that more options on
how to enter focus time can be added there in the future.