communicate errors and some actions through a toast
So far, if an error occurs, the extension will simply do nothing and
users won't be able to understand what's going on unless they look into
the console. The toast now instead displays a prominent error message
when an error occurs, distinguishing between exceptions that the user
can work around and actual errors. This should leave users less puzzled
when things go wrong.

This change required a slight rework of the messages passed between the
extension parts, so that exceptions in the service worker can be passed
on to the UI scripts to react appropriately.

The toast is also used to communicate success of actions that provide no
immediate visual feedback, such as saving options.