let new days start at 4am
When one is working late, after midnight, then the that time should
still be counted towards the previous day. And later that day, users
should be able to start with a clean slate.

There are (presumably fewer) early birds who start working before 4am,
so the switchover time should be configurable.