change behavior on idle and lock states
The previous behavior was difficult: If you'd leave your machine and it
went idle and then auto-locked, then you'd still have to manually leave
your focus block. And if you forget to answer an idle notification, then
no new idle notifications could be issued in the same focus block.

Now, idle notifications are automatically dismissed 30 seconds after the
machine becomes active again. That gives them time to actively respond
to the dialog. If they don't respond, then the event is ignored
entirely, allowing for future idle notifications.

Also, when the system is locked, the focus block will always be
cancelled (possibly retroactively from when the system became idle).
That's less of a problem then it used to be, because if the focus block
should have been continued, then users can manually resume.